refactor(import): generalize preflight missing data selection
