+++ frontend/src/games/edit/EditGamePage.tsx
@@ -11,7 +11,7 @@ import {
 } from '../../api/gameApi';
 import { Game } from '../../database/game';
 import ImportWizard from '../import/ImportWizard';
-import PublishGamePreflight from './PublishGamePreflight';
+import { PublishGamePreflight } from './MissingGameDataPreflight';
 
 interface PreflightData {
     req: CreateGameRequest;
